LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-alignment-implies-size-change-035-expected.txt



PASS .before 1
FAIL .after 2 assert_equals:
<img id="item" src="support/100x100-green.png" class=" before after" data-expected-width="80" data-expected-height="80">
height expected 80 but got 100
